+++ pacman-app-arcade-pacman/src/main/java/de/amr/pacmanfx/arcade/pacman/rendering/ArcadePacMan_HUDRenderer.java
@@ -5,7 +5,7 @@ See file LICENSE in repository root directory for details.
 package de.amr.pacmanfx.arcade.pacman.rendering;
 
 import de.amr.pacmanfx.lib.RectShort;
-import de.amr.pacmanfx.lib.Vector2f;
+import de.amr.pacmanfx.lib.Vector2i;
 import de.amr.pacmanfx.model.Game;
 import de.amr.pacmanfx.model.HUD;
 import de.amr.pacmanfx.model.Score;
@@ -46,7 +46,7 @@ public class ArcadePacMan_HUDRenderer extends BaseSpriteRenderer implements HUDR
     }
 
     @Override
-    public void drawHUD(Game game, HUD hud, Vector2f sceneSize) {
+    public void drawHUD(Game game, HUD hud, Vector2i sceneSize) {
         if (!hud.isVisible()) return;
 
         if (hud.isScoreVisible()) {
