+++ pacman-core/src/main/java/de/amr/pacmanfx/Globals.java
@@ -10,7 +10,6 @@ import de.amr.pacmanfx.controller.GameState;
 import de.amr.pacmanfx.controller.teststates.CutScenesTestState;
 import de.amr.pacmanfx.controller.teststates.LevelMediumTestState;
 import de.amr.pacmanfx.controller.teststates.LevelShortTestState;
-import de.amr.pacmanfx.lib.Vector2f;
 import de.amr.pacmanfx.lib.Vector2i;
 
 import java.util.ArrayList;
@@ -80,7 +79,7 @@ public interface Globals {
      * Arcade maps have a size of 28x36 tiles (including the empty rows over and under the maze). The tile size is 8px
      * which gives a map size of 224x288px.
      */
-    Vector2f ARCADE_MAP_SIZE_IN_PIXELS = new Vector2f(224, 288); // 28x36 tiles
+    Vector2i ARCADE_MAP_SIZE_IN_PIXELS = new Vector2i(224, 288); // 28x36 tiles
 
     /**
      * @return the global object which provides access to different other global objects
