+++ pacman-ui-lib/src/main/java/de/amr/pacmanfx/uilib/rendering/HUDRenderer.java
@@ -5,6 +5,7 @@ See file LICENSE in repository root directory for details.
 package de.amr.pacmanfx.uilib.rendering;
 
 import de.amr.pacmanfx.lib.Vector2f;
+import de.amr.pacmanfx.lib.Vector2i;
 import de.amr.pacmanfx.model.Game;
 import de.amr.pacmanfx.model.HUD;
 
@@ -17,5 +18,5 @@ public interface HUDRenderer extends CanvasRenderer {
      * @param hud the data displayed in the Head-Up Display
      * @param sceneSize scene size in pixels
      */
-    void drawHUD(Game game, HUD hud, Vector2f sceneSize);
+    void drawHUD(Game game, HUD hud, Vector2i sceneSize);
 }
